requirements.txt
numpy==2.1.3 
torch==2.6.0 
openai-whisper==20240930
deepl==1.21.0
pyannote.audio==3.3.2
pydub==0.25.1
noisereduce==3.0.3